[!!!][FEATURE] Make search metadata available
Changes the return format of the execute method to return an array
with a ``nodes`` key containing an array of results. Each result
consists of a ``node`` key with the node instance and a ``meta`` key
containing metadata about the result. Additionally the metadata from
the whole request is available along ``nodes`` as ``response``.

This allows for additional metadata to be passed along, needed for display
of advanced results.

Example of output::

  array(
  	'nodes' => array(
  		array(
  			'node' => NodeInterface,
  			'meta' => array
  		),
  		array(
  			'node' => NodeInterface,
			'meta' => array
  		)
  	),
  	'response' => array
  )

Adjust existing code::

  searchQuery.execute() => searchQuery.execute().nodes

  ${q(node).*} 0> $q(result.node.*}